python/requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0
